Fix use-after-free of frame data in LiveStream
As documented in the destructor for LiveStreamWorker, the head frame
cannot be freed during teardown of the stream, because it may be used
at any time from the GUI thread. It also cannot always be deleted from
the GUI thread, because that would leak in some situations.

The only time when it is safe to delete that frame is during the
destruction of LiveStreamWorker, because the LiveStream has ceased all
use by that time.