Assume that D1-resolution solo6110 video is interlaced
Possibly due to a driver error, libav does not set the frame_interlaced
flag for 6110 frames. Unravelling that driver problem is non-trivial, so
for now, assume all H264 video in D1 resolution (PAL or NTSC) is
interlaced. Worst case, this risks a small decrease in quality from
deinterlacing progressive video.