Disable live view deinterlacing by default
This causes an unacceptable loss of quality on some streams,
particularly on the OSD and around some encoding artifacts. This
could be enabled in the future if we can find or implement a more
useful algorithm; simple blending might even be sufficient.

Part of the problem is that the solo OSD does not seem to be
interlaced, despite being overlaid on an interlaced image. As a
result, it becomes very corrupted when run through a deinterlacer.
That might be fixable with a different algorithm, but we may have
to avoid deinterlacing that particular area of solo images instead.