Hide the live 'Recording' flag until we have RTP support for it
We can't show this if we don't have the information. Previously,
it came from a header added to each multipart segment of the MJPEG
stream (and didn't work in all cases), but there isn't any existing
way to get this information through RTP. We'll have one soon,
probably via a more generic realtime event stream.