Workaround for a MSVC linker bug with libav
With /OPT:REF,ICF (or standard release link options), the MSVC linker
becomes confused and the binary expects to find g_free and g_error_free
defined in avformat-54.dll, not glib-2.0-0.dll, which results in an
unusable binary. Using /DEBUG (as we did before) and not enabling
/OPT:REF,ICF solves the problem at the cost of some binary size.

This can be undone once gstreamer is gone.