Refactor LiveFeedItem and others to always keep a LiveStream
The goal here is to reduce duplication by allowing properties of
the stream to be set directly on that object, even if it hasn't
yet connected. To achieve that, we now create a LiveStream instance
for a camera even if it's offline or we don't have sufficient data
to connect to it; properties can still be set, and it will not
actually connect until it's ready.