Show a statusbar item for server alerts
Display server status alerts as an item in the statusbar, and use
with a yellow indicator for the server. The details of the alert
aren't displayed directly, but clicking on the alert will open the
configuration UI, which should (after #1005) have that information
and ways to act on it.

Feature #1006