Fix LiveStreamItem texture deletion contexts
For the advanced GL case, a variety of paths in LiveStreamItem
would try to delete the texture without any attention to which (if
any) OpenGL context was current, which could result in deleting the
wrong texture (bug #1118) or potentially crashes.

Unfortunately, because there is no link between LiveStreamItem and
the viewport on which it's drawn, it's difficult to cover all of
the situations where textures need to be deleted. This code solves
the problem by checking the current context and setting a flag when
unable to delete, and deleting it later if possible, and careful
checking of the paths where deletion can happen.

There are potentially cleaner (or at least more clearly reliable)
solutions if LiveStreamItem makes assumptions about the scene it's
in, and establishes a direct relationship with the viewport. In
that case, it could detect when the context has been deleted, and
would be able to make the context current where necessary.