Prompt for setup when connecting to unconfigured servers
To assist with the first-time setup of servers, when connecting
to a server with no devices configured, prompt the user via a
dialog to configure that server, launching the configuration UI
if accepted. This should help guide the first time install process
to some extent.

Workaround was necessary to prevent opening two modal windows at
once, which on OS X causes both to freeze up. As a result, the
dialog may come up before the setup wizard has been closed.

Feature #1065