Refactor GstSinkWidget to support non-GL viewports
For systems without proper OpenGL support, it wasn't possible to
play video within the client, due to a hard dependency on GL caused
by inheriting directly from QGLWidget for the GstSinkWidget.

This refactor merges VideoContainer (formerly the parent of a
GstSinkWidget, due to dynamic creation and other needs), and
renders actual video on a viewport child instead. As a result,
GstSinkWidget now paints the frame and other logic, and handles
the gstreamer integration, and the actual painting surface can be
toggled.

Integrated with the live view setting to allow manual enable/disable
of OpenGL. Usual disclaimer with gstreamer stuff applies: this will
all be gone when gstreamer goes.

Bug #1085