Bug #1169: Fix gstreamer freeze/crash during second playback
Reverts commit 67bfcdd8

After playing a video with gstreamer, and when starting playback
of the second video in a session, it seems that the application
stops responding entirely to any events, including input and painting.
Qt on Linux uses glib as the event loop, so it's logical to guess
that gstreamer is breaking the event loop in some way.

Commit 67bfcdd8 introduced asynchronous handling of gstreamer events
on the main event loop, to allow xvideo-related events to be handled.
Removing this and using the synchronous, off-thread handler that is
used for other platforms solves the issue. All platforms use opengl
as a drawing surface, so xvideo event support is no longer used.